Fixed bug that made the leaflet error appear and made some code refactoring
