+++ web/client/product/appConfigEmbedded.js
@@ -7,7 +7,7 @@
 */
 
 import {updateActiveDockEpic, updateMapLayoutEpic} from '../epics/maplayout';
-import {readQueryParamsOnMapEpic, readQueryParamsOnMapInitEpic} from '../epics/queryparams';
+import {readQueryParamsOnMapEpic, switchMapType, checkMapOrientation} from '../epics/queryparams';
 import maplayout from '../reducers/maplayout';
 import searchconfig from '../reducers/searchconfig';
 import version from '../reducers/version';
@@ -58,7 +58,8 @@ export default {
         updateMapLayoutEpic,
         updateActiveDockEpic,
         readQueryParamsOnMapEpic,
-        readQueryParamsOnMapInitEpic
+        switchMapType,
+        checkMapOrientation
     },
     storeOpts: {
         persist: {
